Services/Agro-Chat/requirements.txt
google-generativeai
pillow
streamlit
streamlit-option-menu
googletrans
protobuf
transformers
